+++ rust/cubesql/src/compile/engine/context.rs
@@ -12,9 +12,44 @@ impl SystemVar {
     pub fn new() -> Self {
         let mut variables = HashMap::new();
         variables.insert(
+            "@@max_allowed_packet".to_string(),
+            ScalarValue::UInt32(Some(67108864)),
+        );
+        variables.insert(
+            "@@auto_increment_increment".to_string(),
+            ScalarValue::UInt32(Some(1)),
+        );
+        variables.insert(
             "@@version_comment".to_string(),
             ScalarValue::Utf8(Some("mysql".to_string())),
         );
+        variables.insert(
+            "@@system_time_zone".to_string(),
+            ScalarValue::Utf8(Some("UTC".to_string())),
+        );
+        variables.insert(
+            "@@time_zone".to_string(),
+            ScalarValue::Utf8(Some("SYSTEM".to_string())),
+        );
+        // Isolation old variables
+        variables.insert(
+            "@@tx_isolation".to_string(),
+            ScalarValue::Utf8(Some("REPEATABLE-READ".to_string())),
+        );
+        variables.insert(
+            "@@tx_read_only".to_string(),
+            ScalarValue::Boolean(Some(false)),
+        );
+        // Isolation new variables after 8.0.3
+        variables.insert(
+            "@@transaction_isolation".to_string(),
+            ScalarValue::Utf8(Some("REPEATABLE-READ".to_string())),
+        );
+        variables.insert(
+            "@@transaction_read_only".to_string(),
+            ScalarValue::Boolean(Some(false)),
+        );
+        // Session
         variables.insert(
             "@@sessiontransaction_isolation".to_string(),
             ScalarValue::Utf8(Some("REPEATABLE-READ".to_string())),
@@ -23,6 +58,7 @@ impl SystemVar {
             "@@sessionauto_increment_increment".to_string(),
             ScalarValue::Int64(Some(1)),
         );
+        // character
         variables.insert(
             "@@character_set_client".to_string(),
             ScalarValue::Utf8(Some("utf8mb4".to_string())),
@@ -42,14 +78,6 @@ impl SystemVar {
         variables.insert(
             "@@collation_connection".to_string(),
             ScalarValue::Utf8(Some("utf8mb4_general_ci".to_string())),
-        );
-        variables.insert(
-            "@@system_time_zone".to_string(),
-            ScalarValue::Utf8(Some("UTC".to_string())),
-        );
-        variables.insert(
-            "@@time_zone".to_string(),
-            ScalarValue::Utf8(Some("SYSTEM".to_string())),
         );
 
         Self { variables }